requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
joblib==1.3.1
nltk==3.8.1
protobuf==4.23.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.17
tqdm==4.65.0
typing_extensions==4.7.0
uritemplate==4.1.1
urllib3==1.26.16
